Create SECURITY.md
Security Policy. for the project
Supported Versions
Use this section to tell people about which versions of your project are currently being supported with security updates.

Version	Supported
5.1.x	:white_check_mark:
5.0.x	:x:
4.0.x	:white_check_mark:
< 4.0	:x:
Reporting a Vulnerability
Use this section to tell people how to report a vulnerability.

Tell them where to go, how often they can expect to get an update on a reported vulnerability, what to expect if the vulnerability is accepted or declined, etc.


It seems like you're creating a security policy for a project. This is a great practice to ensure the security and trust of your users. Here's a suggested completion of the security policy:

Supported Versions
Use this section to tell people about which versions of your project are currently being supported with security updates.

Version	Supported
5.1.x	:white_check_mark:
5.0.x	:x:
4.0.x	:white_check_mark:
< 4.0	:x:
This section clearly indicates that versions 5.1.x and 4.0.x are currently supported with security updates, while versions 5.0.x and below 4.0 are not.

Reporting a Vulnerability
If you discover a vulnerability in our project, please report it to security@example.com. We appreciate your help in keeping our project secure.

Response Time

We aim to respond to vulnerability reports within 3 business days. If you don't receive a response within this timeframe, please feel free to send a follow-up email.